Construct slf4j Loggers that also implement LoggingEventAware. (#6)
When Java (or Clojure) code uses a LoggingEventBuilder like

    logger.atDebug().setMessage("Temperature changed.").addKeyValue("oldT", oldT).addKeyValue("newT", newT).log();

Then the default implementation will do this:
<https://github.com/qos-ch/slf4j/blob/v_2.0.16/slf4j-api/src/main/java/org/slf4j/spi/DefaultLoggingEventBuilder.java#L163>. So
if the Logger implements org.slf4j.spi.LoggingEventAware, it gets to
handle the LoggingEvent's key-value pairs in a first-class manner. But
if it doesn't, that data is just shoved into the message. It's
preferable to get key-value pairs and log them as a JSON map, so
construct Logger objects that implement LoggingEventAware and do that.